updated the post component, scrollbar styles
