+++ Frontend/src/components/CommunitySidebar.vue
@@ -2,7 +2,6 @@
     <div class="w-60 flex flex-col gap-4 select-none border-r border-[#242c2e]
                  sticky top-16 overflow-hidden hover:overflow-y-scroll h-[90vh]
                  scrollbar-styled">
-
         <div>
             <div v-for="n in 3" :key="n">
                 <RowComponent :name="subreddits[n]"/>
