+++ Frontend/src/views/HomeView.vue
@@ -11,11 +11,12 @@ import IconChevronUp from "@/components/icons/IconChevronUp.vue";
         <PostFeed/>
         <div class="
         w-1/3 flex flex-col gap-4
-        sticky top-16 overflow-hidden hover:overflow-y-scroll h-fit max-h-[90vh]
+        sticky top-16 overflow-hidden hover:overflow-y-auto h-fit max-h-[90vh]
         bg-black rounded-lg p-4 scrollbar-styled">
-
-            <p>Recent posts</p>
-            <p class="hover:text-blue-600 cursor-pointer">Clear</p>
+            <div class="flex justify-between">
+                <p>Recent posts</p>
+                <p class="hover:text-blue-600 cursor-pointer">Clear</p>
+            </div>
 
         </div>
     </div>
